fastapi-ai/requirements.txt
aioice==0.10.1
aiortc==1.14.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
av==16.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Cython==3.2.3
dnspython==2.8.0
easydict==1.13
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
fonttools==4.61.1
google-crc32c==1.7.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
ImageIO==2.37.2
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
msgpack==1.1.2
networkx==3.6.1
numpy==2.2.6
onnx==1.20.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.4
packaging==25.0
passlib==1.7.4
pgvector==0.4.2
pillow==12.0.0
prettytable==3.17.0
protobuf==6.33.2
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.8.0
scipy==1.16.3
sentry-sdk==2.44.0
shellingham==1.5.4
simsimd==6.5.8
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.49.3
stringzilla==4.5.1
threadpoolctl==3.6.0
tifffile==2025.12.12
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
